ResourceRow.ts moved to utils folder (updated imports in many files) | ResourcePage added minimal text for vars like Address: ...
